ml-bridge-py/pyproject.toml
[project]
name = "ml-bridge-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "absl-py>=2.1.0",
    "iree-base-compiler>=3.1.0",
    "iree-base-runtime>=3.1.0",
    "jax>=0.5.0",
    "pytest>=8.3.4",
]

[tool.pytest.ini_options]
testpaths = ["test.py"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--maxfail=1 --disable-warnings"
